// client/src/app/core/Models/book.model.ts
export class Book {
  _id: string = '';
  title: string = '';
  genre: string = '';
  author: string = '';
  pages: number = 0;
  rating: number = 0;
  price: number = 0;
  favorite: boolean = false;
  archive: boolean = false;
  desc: string = '';
  createdTime: number = 0;
  lastViewedTime: number = 0;
}